link to web4 from bos
e2e tests run on static http server like web4
