+++ bos_components/account_report.jsx
@@ -1,1 +1,1 @@
-return <iframe style={{ width: "100%", height: "600px" }} src="http://localhost:8081"></iframe>
+return <a href="https://arizportfolio.near.page">arizportfolio.near.page</a>
