+++ playwright.config.js
@@ -38,24 +38,10 @@ export default defineConfig({
     /* Maximum time each action such as `click()` can take. Defaults to 0 (no limit). */
     actionTimeout: 0,
     /* Base URL to use in actions like `await page.goto('/')`. */
-    baseURL: "http://localhost:8080",
+    baseURL: "http://localhost:8081",
 
     /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
     trace: "on-first-retry",
-    storageState: {
-      cookies: [],
-      origins: [
-        {
-          origin: "http://localhost:8080",
-          localStorage: [
-            {
-              name: "flags",
-              value: JSON.stringify({ bosLoaderUrl: "http://127.0.0.1:3030" }),
-            },
-          ],
-        },
-      ],
-    },
   },
 
   /* Configure projects for major browsers */
